Refer to note #1.1

+++ WeAct_MiniSTM32_F401/src/LEDs/startup.c
@@ -19,6 +19,7 @@ extern void main(void);
 
 __attribute__((interrupt, naked, noreturn)) void RESET(void)
 {
+
 	//Disable all interrupts
 	Disable_System_Interrupts();
 	Disable_External_Interrupts();
@@ -39,19 +40,20 @@ __attribute__((interrupt, naked, noreturn)) void RESET(void)
 	//84MHz High speed APB2
 	__maxout_system_domains();
 
+
 	//Enable A, B, and C GPIO ports clock
 	ENABLE_AHB_PORT_A();
 	ENABLE_AHB_PORT_B();
 	ENABLE_AHB_PORT_C();
 
 	//Copy the data at the end of text section to RAM where it should be accessed
-	for(uint32_t *t_pos = &__etext__, *d_pos = &__sdata__; d_pos < &__edata__; d_pos++, t_pos++)
+	for(uint8_t *t_pos = &__etext__, *d_pos = &__sdata__; d_pos < &__edata__; d_pos++, t_pos++)
 	{
 		*d_pos = *t_pos;
 	}
 
 	//Init the bss section to zero
-	for(uint32_t *b_pos = &__sbss__; b_pos < &__ebss__; b_pos++)
+	for(uint8_t *b_pos = &__sbss__; b_pos < &__ebss__; b_pos++)
 	{
 		*b_pos = 0UL;
 	}
